+++ src/main/org/firebirdsql/gds/impl/jni/EmbeddedGDSFactoryPlugin.java
@@ -29,7 +29,7 @@ public class EmbeddedGDSFactoryPlugin extends BaseGDSFactoryPlugin {
     private static final String[] TYPE_ALIASES = new String[0];
 
     private static final String[] JDBC_PROTOCOLS = new String[] { 
-        "jdbc:firebirdsql:embedded:"
+        "jdbc:firebirdsql:embedded:", "jdbc:firebird:embedded:"
     };
 
     public String getPluginName() {
@@ -48,8 +48,11 @@ public class EmbeddedGDSFactoryPlugin extends BaseGDSFactoryPlugin {
         return JDBC_PROTOCOLS;
     }
     
-    public String getDatabasePath(String server, Integer port, String path)
-            throws GDSException {
+    public String getDatabasePath(String server, Integer port, String path) throws GDSException {
+        if (path == null) {
+            throw new GDSException("Database name/path is required.");
+        }
+
         return path;
     }