+++ src/main/org/firebirdsql/gds/impl/jni/NativeGDSFactoryPlugin.java
@@ -26,11 +26,13 @@ public class NativeGDSFactoryPlugin extends BaseGDSFactoryPlugin {
 
     public static final String NATIVE_TYPE_NAME = "NATIVE";
 
-    private static final String[] TYPE_ALIASES = new String[]{"TYPE2"};
-    private static final String[] JDBC_PROTOCOLS = new String[]{"jdbc:firebirdsql:native:"};
+    private static final String[] TYPE_ALIASES = new String[] { "TYPE2" };
+    private static final String[] JDBC_PROTOCOLS = new String[] {
+            "jdbc:firebirdsql:native:", "jdbc:firebird:native:"
+    };
     
     public String getPluginName() {
-        return "JNI-based GDS implementation.";
+        return "JNA-based GDS implementation.";
     }
 
     public String getTypeName() {
@@ -46,18 +48,19 @@ public class NativeGDSFactoryPlugin extends BaseGDSFactoryPlugin {
     }
 
     public String getDatabasePath(String server, Integer port, String path) throws GDSException{
-        if (server == null)
-            throw new GDSException("Server name/address is required " +
-                    "for native implementation.");
-        
-        if (path == null)
+        if (server == null) {
+            throw new GDSException("Server name/address is required for native implementation.");
+        }
+        if (path == null) {
             throw new GDSException("Database name/path is required.");
+        }
         
         StringBuilder sb = new StringBuilder();
         
         sb.append(server);
-        if (port != null)
+        if (port != null) {
             sb.append('/').append(port.intValue());
+        }
         
         sb.append(':').append(path);